backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-dotenv
requests
django-storages
boto3
django-environ
django-cors-headers
